extracted hooks from create flight page
